Renames all .css files in the src/components/Grid to .scss. Reason is monorepo is not setup for css modules on .css files, and, I have used composes: in some of these and they're not working when monorepo tries to consume
