+++ src/components/Grid/Pagination.md
@@ -8,7 +8,7 @@ import { Grid } from '../Grid/Grid.js'
 import { Row } from '../Grid/Row.js'
 import { Column } from '../Grid/Column.js'
 import { useGridSorting } from '../Grid/useGridSorting.js'
-import styles from '../Grid/grid-example.module.css'
+import styles from '../Grid/grid-example.module.scss'
 
 const pageOne = `
 {"page":1,"per_page":6,"total":12,"total_pages":2,"data":[{"id":1,"name":"cerulean","year":2000,"color":"#98B2D1","pantone_value":"15-4020"},{"id":2,"name":"fuchsia rose","year":2001,"color":"#C74375","pantone_value":"17-2031"},{"id":3,"name":"true red","year":2002,"color":"#BF1932","pantone_value":"19-1664"},{"id":4,"name":"aqua sky","year":2003,"color":"#7BC4C4","pantone_value":"14-4811"},{"id":5,"name":"tigerlily","year":2004,"color":"#E2583E","pantone_value":"17-1456"},{"id":6,"name":"blue turquoise","year":2005,"color":"#53B0AE","pantone_value":"15-5217"}]}
