+++ src/components/Grid/useGridSorting.module.scss
@@ -1,6 +1,6 @@
 /**
  * Sort icons UX (from default, up, to down) seem to follow the same
- * conventions as Bootsprap: https://examples.bootstrap-table.com/
+ * conventions as Bootstrap: https:\/\/examples.bootstrap-table.com/
  */
 .icon {
   margin-left: var(--Space-4);
