+++ src/components/Grid/usePagination.js
@@ -1,5 +1,5 @@
 import React, { useEffect, useState } from 'react'
-import styles from './Pagination.module.css'
+import styles from './Pagination.module.scss'
 
 const usePagination = ({
   data = null,
